move bulk editing code to stimulus
