+++ app/javascript/application.ts
@@ -8,7 +8,6 @@ import 'masonry-layout'
 
 import Cocooned from '@notus.sh/cocooned'
 
-import 'src/bulk_edit'
 import 'src/passwords'
 
 // Include stimulus controllers
