--- app/javascript/src/bulk_edit.ts
@@ -1,60 +1,0 @@
-function updateTagOptions (tags: string[], input, addTags = true): void {
-  tags.forEach((tag) => {
-    if (addTags) {
-      input.tomselect.addOption({ value: tag, text: tag })
-    } else {
-      input.tomselect.removeOption({ value: tag, text: tag })
-    }
-  })
-  input.tomselect.refreshOptions(false)
-}
-
-function getTags (modelId: string): string[] {
-  const selector = `[data-bulk-item-tags="${modelId}"]`
-  const tagLinks = document.querySelectorAll(
-    selector
-  )
-  return Array.prototype.slice
-    .call(tagLinks)
-    .map((tag: HTMLAnchorElement) => tag.textContent)
-}
-
-function updateTagList (modelId: string, add: boolean): void {
-  const tags = getTags(modelId)
-  if (tags.length > 0) {
-    updateTagOptions(tags, document.querySelector('select[name="remove_tags[]"]'), add)
-  }
-}
-
-function handleCheckboxChange (event): void {
-  const target = event.target as HTMLInputElement
-  event.preventDefault()
-  // the bulk select checkbox has been selected
-  if (target.name === 'bulk-select-all') {
-    document
-      .querySelectorAll('[data-bulk-item]')
-      .forEach((checkbox: HTMLInputElement) => {
-        const modelId = checkbox.getAttribute('data-bulk-item')
-        if (modelId != null) {
-          updateTagList(modelId, target.checked)
-        }
-        checkbox.checked = target.checked
-      })
-  } else {
-    // a single checkbox item has been selected.
-    const modelId = target.getAttribute('data-bulk-item') as string
-    if (modelId != null) {
-      updateTagList(modelId, target.checked)
-    }
-  }
-}
-
-document.addEventListener('ManyfoldReady', () => {
-  const bulkEditTable = document.querySelector(
-    '[data-bulk-edit]'
-  )
-  if (bulkEditTable != null) {
-    bulkEditTable.removeEventListener('change', handleCheckboxChange)
-    bulkEditTable.addEventListener('change', handleCheckboxChange)
-  }
-})
