+++ app/views/models/bulk_edit.html.erb
@@ -8,7 +8,7 @@
 
 <%= form_with url: update_models_path, method: :patch do |form| %>
 
-  <table class="table table-striped" data-bulk-edit>
+  <table class="table table-striped" data-controller="bulk-edit" data-action="change->bulk-edit#handleCheckboxChange">
     <tr>
       <th><input type="checkbox" aria-label="<%= translate ".select_all" %>" value="0" name="bulk-select-all"></th>
       <th><%= Model.human_attribute_name(:name) %></th>
