+++ app/views/problems/index.html.erb
@@ -11,7 +11,7 @@
 <% end %>
 
 <%= form_with url: resolve_problems_path, method: :post, class: "mb-3", data: {confirm: translate(".bulk_confirm")} do |form| %>
-  <table class="table table-striped" data-bulk-edit>
+  <table class="table table-striped" data-controller="bulk-edit" data-action="change->bulk-edit#handleCheckboxChange">
     <tr>
       <th><input type="checkbox" aria-label="<%= translate ".select_all" %>" value="0" name="bulk-select-all"></th>
       <th colspan=2><%= t(".item") %></th>
